:recycle: [repositories] Move function `descend` to `repository`
